expression: change boolean expression parser
The specification at:

  http://www.opengroup.org/onlinepubs/009695399/utilities/find.html

is quite confusing w.r.t. precedence of operators. It is unclear whether
a left-to-right parsing should be done, i.e.:

  a AND b OR c <=> a AND (b OR c)

or if one should follow the mathematical definition, i.e.:

  a AND b OR c <=> (a AND b) OR c

I implemented the first one. However, it is totally counter intuitive.
Besides, the BSD implementation follows the second.

This patch discards the first behavior in favor of the latter.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>